judge: optimize
